Implemented Logout: - Created standalone navbar component, added logout button - Implemented logout action, updated reducer - Destroy local storage token on logout action - Modularized action creator exports
