+++ hab/src/App.js
@@ -1,18 +1,18 @@
 import React, { useEffect }from 'react';
-import { Navbar, NavbarBrand } from 'shards-react';
 import { BrowserRouter as Router, Route, Switch } from 'react-router-dom';
 
 import Login from './containers/Login';
 import Register from './containers/Register';
 import Home from './components/Home';
 import Loading from './components/Loading';
+import Navbar from './components/Navbar';
 import { useStateValue } from './state';
-import { useReturnAction } from './actions/Return';
-
+import { useReturnAction, useLogoutAction } from './actions';
 
 function App() {
   const [{ userName, token, loading }] = useStateValue();
   const { returnAction } = useReturnAction();
+  const { logoutAction } = useLogoutAction();
   const localToken = localStorage.getItem('token', token);
 
   useEffect(() => {
@@ -22,9 +22,7 @@ function App() {
 
   return (
       <div className="App">
-        <Navbar type="dark" theme="primary" expand="md" className="mb-2">
-          <NavbarBrand href="#">React Boilerplate</NavbarBrand> 
-        </Navbar>
+        <Navbar userName={userName} logout={logoutAction} />
         <Router>
           <Switch>
             <Route path="/register">
